requirements.txt
alembic==1.13.1
APScheduler
bcrypt==4.1.2
httpx==0.27.0
jsonref==0.2
Mako==1.3.2
marshmallow==3.21.1
marshmallow_enum==1.5.1
PyYAML==6.0.1
pydantic==2.6.4
python-dateutil==2.9.0.post0
sanic==22.12.0
Sanic-Cors==2.2.0
asyncpg==0.29.0
SQLAlchemy==2.0.29
contextvars==2.4
sqlalchemy-tx-context==0.2.0
qstd-async-tools==0.1.0
qstd-config==0.2.0
qstd-logger-json-formatter==0.1.0
qstd_core==0.1.0